requirements.txt
Django
psycopg2-binary
python-dotenv
djangorestframework
django-cors-headers
gunicorn
dj-database-url
whitenoise
cloudinary
django-cloudinary-storage
Pillow
djangorestframework-simplejwt
